requirements.txt
python-telegram-bot==13.13
beautifulsoup4==4.11.1
requests==2.28.1